Fixed problems with exporting to jar
There was an issue with the jar file properly getting file resources to read. I changed each line to call getResourceAsStream instead of a raw path. There was a minor problem with the export renaming the movies text file because it is the same name as the package. Moved the file to its own package too. Jar file is working properly now.